requirements.txt
django==5.0.0
djgentelella==0.3.21
djangorestframework==3.15.2
